requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.13
    # via -r requirements.in
aiosignal==1.3.2
    # via aiohttp
altair==4.2.2
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
attrs==25.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   outcome
    #   referencing
    #   trio
beautifulsoup4==4.13.4
    # via -r requirements.in
blinker==1.9.0
    # via streamlit
cachetools==6.1.0
    # via streamlit
certifi==2025.6.15
    # via
    #   httpcore
    #   httpx
    #   requests
    #   selenium
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   nltk
    #   streamlit
distro==1.9.0
    # via openai
entrypoints==0.4
    # via altair
faiss-cpu==1.11.0
    # via -r requirements.in
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
h11==0.16.0
    # via
    #   httpcore
    #   wsproto
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   trio
    #   yarl
importlib-metadata==8.7.0
    # via streamlit
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jiter==0.10.0
    # via openai
joblib==1.5.1
    # via nltk
jsonschema==4.24.0
    # via altair
jsonschema-specifications==2025.4.1
    # via jsonschema
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
multidict==6.4.4
    # via
    #   aiohttp
    #   yarl
nltk==3.9.1
    # via -r requirements.in
numpy==2.3.0
    # via
    #   -r requirements.in
    #   altair
    #   faiss-cpu
    #   pandas
    #   pydeck
    #   streamlit
openai==1.88.0
    # via -r requirements.in
outcome==1.3.0.post0
    # via
    #   trio
    #   trio-websocket
packaging==25.0
    # via
    #   faiss-cpu
    #   streamlit
    #   webdriver-manager
pandas==2.3.0
    # via
    #   altair
    #   streamlit
pillow==11.2.1
    # via
    #   -r requirements.in
    #   streamlit
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
protobuf==3.20.3
    # via streamlit
pyarrow==20.0.0
    # via streamlit
pydantic==2.11.7
    # via openai
pydantic-core==2.33.2
    # via pydantic
pydeck==0.9.1
    # via streamlit
pygments==2.19.1
    # via rich
pympler==1.1
    # via streamlit
pysocks==1.7.1
    # via urllib3
python-dateutil==2.9.0.post0
    # via
    #   pandas
    #   streamlit
python-dotenv==1.1.0
    # via webdriver-manager
pytz==2025.2
    # via pandas
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via
    #   nltk
    #   tiktoken
requests==2.32.4
    # via
    #   -r requirements.in
    #   streamlit
    #   tiktoken
    #   webdriver-manager
rich==14.0.0
    # via streamlit
rpds-py==0.25.1
    # via
    #   jsonschema
    #   referencing
selenium==4.33.0
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   openai
    #   trio
sortedcontainers==2.4.0
    # via trio
soupsieve==2.7
    # via beautifulsoup4
streamlit==1.22.0
    # via -r requirements.in
tenacity==8.5.0
    # via streamlit
tiktoken==0.9.0
    # via -r requirements.in
toml==0.10.2
    # via streamlit
toolz==1.0.0
    # via altair
tornado==6.5.1
    # via streamlit
tqdm==4.67.1
    # via
    #   -r requirements.in
    #   nltk
    #   openai
trio==0.30.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.12.2
    # via selenium
typing-extensions==4.13.2
    # via
    #   anyio
    #   beautifulsoup4
    #   openai
    #   pydantic
    #   pydantic-core
    #   referencing
    #   selenium
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
tzlocal==5.3.1
    # via streamlit
urllib3[socks]==2.4.0
    # via
    #   requests
    #   selenium
validators==0.35.0
    # via streamlit
webdriver-manager==4.0.2
    # via -r requirements.in
websocket-client==1.8.0
    # via selenium
wsproto==1.2.0
    # via trio-websocket
yarl==1.20.1
    # via aiohttp
zipp==3.23.0
    # via importlib-metadata
